// assets/src/__generated__/ProductCommunityItemsUpdateProductAnswerMutation.graphql.ts
export type CommunityModerationStatus = "HIDDEN" | "PENDING" | "PUBLISHED" | "REJECTED" | "REMOVED" | "%future added value";
export type UpdateProductAnswerInput = {
  body?: string | null | undefined;
  id: string;
};
export type ProductCommunityItemsUpdateProductAnswerMutation$variables = {
  input: UpdateProductAnswerInput;
};
export type ProductCommunityItemsUpdateProductAnswerMutation$data = {
  readonly updateProductAnswer: {
    readonly answer: {
      readonly body: string;
      readonly id: string;
      readonly moderationStatus: CommunityModerationStatus;
    } | null | undefined;
    readonly errors: ReadonlyArray<{
      readonly code: string;
      readonly field: string | null | undefined;
      readonly message: string;
    }>;
  };
};
export type ProductCommunityItemsUpdateProductAnswerMutation = {
  response: ProductCommunityItemsUpdateProductAnswerMutation$data;
  variables: ProductCommunityItemsUpdateProductAnswerMutation$variables;
};