// assets/src/__generated__/ProductCommunityItemsUpdateProductReviewMutation.graphql.ts
export type CommunityModerationStatus = "HIDDEN" | "PENDING" | "PUBLISHED" | "REJECTED" | "REMOVED" | "%future added value";
export type UpdateProductReviewInput = {
  body?: string | null | undefined;
  id: string;
  rating?: number | null | undefined;
  title?: string | null | undefined;
};
export type ProductCommunityItemsUpdateProductReviewMutation$variables = {
  input: UpdateProductReviewInput;
};
export type ProductCommunityItemsUpdateProductReviewMutation$data = {
  readonly updateProductReview: {
    readonly errors: ReadonlyArray<{
      readonly code: string;
      readonly field: string | null | undefined;
      readonly message: string;
    }>;
    readonly review: {
      readonly body: string | null | undefined;
      readonly id: string;
      readonly moderationStatus: CommunityModerationStatus;
      readonly rating: number;
      readonly title: string | null | undefined;
    } | null | undefined;
  };
};
export type ProductCommunityItemsUpdateProductReviewMutation = {
  response: ProductCommunityItemsUpdateProductReviewMutation$data;
  variables: ProductCommunityItemsUpdateProductReviewMutation$variables;
};